requirements.txt
numpy==1.19.0
pyglet==1.5.0
gym-retro==0.8.0
tensorflow==2.3.1
